[UX] moved login from root level to platform subcommand
